# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  contracts.FilePreview:
    properties:
      download_path:
        type: string
      internal_url:
        type: string
      media_type:
        type: string
      modified:
        type: string
      name:
        type: string
      path:
        type: string
      size:
        type: integer
    type: object
...
  contracts.PreviewSummary:
    properties:
      movie_files:
        type: integer
      other_files:
        type: integer
      total_files:
        type: integer
      total_size:
        type: string
      tv_files:
        type: integer
      video_files:
        type: integer
    type: object
  contracts.QuickTaskRequest:
    properties:
      created_by:
        type: integer
      path:
        type: string
      type:
        enum:
        - daily
        - recent
        - weekly
        - realtime
        type: string
    required:
    - created_by
    - type
    type: object
...
  contracts.TaskListResponse:
    properties:
      summary:
        $ref: '#/definitions/contracts.TaskSummary'
      tasks:
        items:
          $ref: '#/definitions/contracts.TaskResponse'
        type: array
      total_count:
        type: integer
    type: object
...
  contracts.TaskPreviewResponse:
    properties:
      files:
        items:
          $ref: '#/definitions/contracts.FilePreview'
        type: array
      summary:
        $ref: '#/definitions/contracts.PreviewSummary'
      task:
        $ref: '#/definitions/contracts.TaskResponse'
      time_range:
        $ref: '#/definitions/contracts.TimeRange'
    type: object
  contracts.TaskRequest:
    properties:
      auto_preview:
        type: boolean
      created_by:
        type: integer
      cron_expr:
        type: string
      enabled:
        type: boolean
      hours_ago:
        description: 最多1年
        maximum: 8760
        minimum: 1
        type: integer
      name:
        maxLength: 100
        minLength: 1
        type: string
      path:
        type: string
      video_only:
        type: boolean
    required:
    - cron_expr
    - hours_ago
    - name
    - path
    type: object
  contracts.TaskResponse:
    properties:
      auto_preview:
        type: boolean
      created_at:
        type: string
      created_by:
        type: integer
      cron_expr:
        type: string
      enabled:
        type: boolean
      failure_count:
        type: integer
      hours_ago:
        type: integer
      id:
        type: string
      last_run_at:
        type: string
      name:
        type: string
      next_run_at:
        type: string
      path:
        type: string
      run_count:
        type: integer
      status:
        $ref: '#/definitions/entities.TaskStatus'
      success_count:
        type: integer
      updated_at:
        type: string
      video_only:
        type: boolean
    type: object
  contracts.TaskRunRequest:
    properties:
      force_run:
        type: boolean
      notify_user:
        type: boolean
      preview:
        type: boolean
      task_id:
        type: string
    required:
    - task_id
    type: object
  contracts.TaskRunResponse:
    properties:
      download_ids:
        items:
          type: string
        type: array
      preview:
        $ref: '#/definitions/contracts.TaskPreviewResponse'
      run_id:
        type: string
      started_at:
        type: string
      status:
        type: string
      task_id:
        type: string
    type: object
  contracts.TaskSummary:
    properties:
      disabled_count:
        type: integer
      enabled_count:
        type: integer
      error_count:
        type: integer
      running_count:
        type: integer
    type: object
  contracts.TaskUpdateRequest:
    properties:
      auto_preview:
        type: boolean
      cron_expr:
        type: string
      enabled:
        type: boolean
      hours_ago:
        maximum: 8760
        minimum: 1
        type: integer
      name:
        maxLength: 100
        minLength: 1
        type: string
      path:
        type: string
      video_only:
        type: boolean
    type: object
  contracts.TimeRange:
    properties:
      end:
        type: string
      start:
        type: string
    type: object
...
  entities.TaskStatus:
    enum:
    - idle
    - running
    - success
    - error
    - stopped
    type: string
    x-enum-comments:
      TaskStatusError: 最后一次执行失败
      TaskStatusIdle: 空闲状态
      TaskStatusRunning: 运行中
      TaskStatusStopped: 已停止
      TaskStatusSuccess: 最后一次执行成功
    x-enum-descriptions:
    - 空闲状态
    - 运行中
    - 最后一次执行成功
    - 最后一次执行失败
    - 已停止
    x-enum-varnames:
    - TaskStatusIdle
    - TaskStatusRunning
    - TaskStatusSuccess
    - TaskStatusError
    - TaskStatusStopped
...
host: localhost:8081
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: 基于Gin框架的Alist和Aria2下载管理服务
  license:
    name: MIT
    url: https://github.com/easayliu/alist-aria2-download/blob/main/LICENSE
  termsOfService: http://swagger.io/terms/
  title: Alist Aria2 Download API
  version: "1.0"
paths:
...
  /tasks:
    get:
      consumes:
      - application/json
      description: 获取所有定时任务的列表
      parameters:
      - description: 创建者ID
        in: query
        name: created_by
        type: integer
      - description: 是否启用
        in: query
        name: enabled
        type: boolean
      - description: 任务状态
        in: query
        name: status
        type: string
      - default: 100
        description: 限制数量
        in: query
        name: limit
        type: integer
      - default: 0
        description: 偏移量
        in: query
        name: offset
        type: integer
      - description: 排序字段
        in: query
        name: sort_by
        type: string
      - description: 排序方向
        in: query
        name: sort_order
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 任务列表
          schema:
            $ref: '#/definitions/contracts.TaskListResponse'
        "500":
          description: 服务器内部错误
          schema:
            additionalProperties: true
            type: object
      summary: 获取定时任务列表
      tags:
      - 定时任务
    post:
      consumes:
      - application/json
      description: 创建一个新的定时任务，按照cron表达式定期执行
      parameters:
      - description: 创建任务请求
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/contracts.TaskRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 任务创建成功
          schema:
            $ref: '#/definitions/contracts.TaskResponse'
        "400":
          description: 请求参数错误
          schema:
            additionalProperties: true
            type: object
        "500":
          description: 服务器内部错误
          schema:
            additionalProperties: true
            type: object
      summary: 创建定时任务
      tags:
      - 定时任务
  /tasks/{id}:
    delete:
      consumes:
      - application/json
      description: 删除指定ID的定时任务
      parameters:
      - description: 任务ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 任务删除成功
          schema:
            additionalProperties: true
            type: object
        "404":
          description: 任务不存在
          schema:
            additionalProperties: true
            type: object
        "500":
          description: 服务器内部错误
          schema:
            additionalProperties: true
            type: object
      summary: 删除定时任务
      tags:
      - 定时任务
    get:
      consumes:
      - application/json
      description: 根据任务ID获取定时任务的详细信息
      parameters:
      - description: 任务ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 任务详情
          schema:
            $ref: '#/definitions/contracts.TaskResponse'
        "404":
          description: 任务不存在
          schema:
            additionalProperties: true
            type: object
        "500":
          description: 服务器内部错误
          schema:
            additionalProperties: true
            type: object
      summary: 获取定时任务详情
      tags:
      - 定时任务
    put:
      consumes:
      - application/json
      description: 更新指定ID的定时任务信息
      parameters:
      - description: 任务ID
        in: path
        name: id
        required: true
        type: string
      - description: 更新任务请求
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/contracts.TaskUpdateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 任务更新成功
          schema:
            $ref: '#/definitions/contracts.TaskResponse'
        "400":
          description: 请求参数错误
          schema:
            additionalProperties: true
            type: object
        "404":
          description: 任务不存在
          schema:
            additionalProperties: true
            type: object
        "500":
          description: 服务器内部错误
          schema:
            additionalProperties: true
            type: object
      summary: 更新定时任务
      tags:
      - 定时任务
  /tasks/{id}/disable:
    post:
      consumes:
      - application/json
      description: 禁用指定ID的定时任务
      parameters:
      - description: 任务ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 任务禁用成功
          schema:
            additionalProperties: true
            type: object
        "404":
          description: 任务不存在
          schema:
            additionalProperties: true
            type: object
        "500":
          description: 服务器内部错误
          schema:
            additionalProperties: true
            type: object
      summary: 禁用定时任务
      tags:
      - 定时任务
  /tasks/{id}/enable:
    post:
      consumes:
      - application/json
      description: 启用指定ID的定时任务
      parameters:
      - description: 任务ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 任务启用成功
          schema:
            additionalProperties: true
            type: object
        "404":
          description: 任务不存在
          schema:
            additionalProperties: true
            type: object
        "500":
          description: 服务器内部错误
          schema:
            additionalProperties: true
            type: object
      summary: 启用定时任务
      tags:
      - 定时任务
  /tasks/{id}/preview:
    get:
      consumes:
      - application/json
      description: 预览定时任务将要下载的文件，不实际执行下载
      parameters:
      - description: 任务ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 预览结果
          schema:
            $ref: '#/definitions/contracts.TaskPreviewResponse'
        "404":
          description: 任务不存在
          schema:
            additionalProperties: true
            type: object
        "500":
          description: 服务器内部错误
          schema:
            additionalProperties: true
            type: object
      summary: 预览定时任务
      tags:
      - 定时任务
  /tasks/{id}/run:
    post:
      consumes:
      - application/json
      description: 立即执行指定ID的定时任务，不等待下一个调度时间
      parameters:
      - description: 任务ID
        in: path
        name: id
        required: true
        type: string
      - description: 执行选项
        in: body
        name: request
        schema:
          $ref: '#/definitions/contracts.TaskRunRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 任务已开始执行或预览结果
          schema:
            $ref: '#/definitions/contracts.TaskRunResponse'
        "404":
          description: 任务不存在
          schema:
            additionalProperties: true
            type: object
        "500":
          description: 服务器内部错误
          schema:
            additionalProperties: true
            type: object
      summary: 立即执行定时任务
      tags:
      - 定时任务
  /tasks/{id}/stop:
    post:
      description: 停止正在运行的任务
      parameters:
      - description: 任务ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 任务停止成功
          schema:
            additionalProperties: true
            type: object
        "400":
          description: 请求参数错误
          schema:
            additionalProperties: true
            type: object
        "500":
          description: 服务器内部错误
          schema:
            additionalProperties: true
            type: object
      summary: 停止任务
      tags:
      - 定时任务
  /tasks/quick:
    post:
      consumes:
      - application/json
      description: 创建预定义的快捷任务（每日、最近等）
      parameters:
      - description: 快捷任务请求
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/contracts.QuickTaskRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 任务创建成功
          schema:
            $ref: '#/definitions/contracts.TaskResponse'
        "400":
          description: 请求参数错误
          schema:
            additionalProperties: true
            type: object
        "500":
          description: 服务器内部错误
          schema:
            additionalProperties: true
            type: object
      summary: 创建快捷任务
      tags:
      - 定时任务
  /tasks/scheduler/status:
    get:
      description: 获取任务调度器的状态信息
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: 获取调度器状态
      tags:
      - 定时任务
  /tasks/statistics:
    get:
      description: 获取任务系统的统计信息
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: 获取任务统计
      tags:
      - 定时任务
  /tasks/user/{user-id}:
    get:
      description: 获取指定用户的所有任务列表
      parameters:
      - description: 用户ID
        in: path
        name: user-id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 用户任务列表
          schema:
            additionalProperties: true
            type: object
        "400":
          description: 请求参数错误
          schema:
            additionalProperties: true
            type: object
        "500":
          description: 服务器内部错误
          schema:
            additionalProperties: true
            type: object
      summary: 获取用户任务
      tags:
      - 定时任务
schemes:
- http
- https
swagger: "2.0"